frontend/authority/requirements.txt
flask==3.0.3
requests==2.32.3
dnspython==2.6.1
gunicorn==22.0.0